// frontend/src/app/shared/interfaces/enuns/header-type.enum.ts
/**
 * Type of header (Authorization token, CORS, etc)
 */
export enum HeaderType {
  CORS = 'CORS',
  ACCEPT_JSON = 'ACCEPT_JSON',
  CONTENT_JSON = 'CONTENT_JSON',
  CONTENT_MULTIPART = 'CONTENT_MULTIPART',
  AUTH = 'AUTH',
  ALLOW_ALL_METHODS = 'ALLOW_ALL_METHODS',
  ALLOW_COMMON_METHODS = 'ALLOW_COMMON_METHODS',
  AUTHORIZATION = 'AUTHORIZATION',
}